+++ docs/.vitepress/config.ts
@@ -49,6 +49,7 @@ export default withMermaid(
         items: [
           { text: "What is Kinetix?", link: "/" },
           { text: "Getting Started", link: "/installation" },
+          { text: "Component previews", link: "/component-previews" },
         ],
       },
       {